chore(tests): workaround for lune bug without using `sh` pipes
When a command such as `unzip` is spawned, a pipe is established between
lune's stdin and and the `unzip`'s stdout. The pipe buffer has a fixed size
which must not be overflowed, and hence processes yield writing to it, and
wait for the consumer to read the data being written.

`process.create` and `process.exec` wait for the process to exit first, before
calling `read_to_end` on the stream, which works for processes with small outputs,
but for larger ones such as `unzip`'s which involve the process itself waiting for
consumption of data, it creates a cyclic yielding dependence where lune waits for
the process to exit, which wouldn't happen until it finishes writing its data, which
in turn would only happen if it can continue writing chunks after the previous data
has been read (which lune is waiting for the process to exit to begin doing).

We work around this by using a Luau implementation of `ChildReader:readToEnd`.
Specifically, it is backed by a table in order to prevent expensive hashing required
for string concatenation. Another workaround is to compare data as it is being streamed
down to us in real-time, which is only applicable for a few test cases.