refactor: update to new luau require-by-string semantics (#3)
* feat: upgrade to new require by string semantics

* fix: fix tests as good as possible

* chore(nix): update configuration

* chore(nix): use github release of lune

* chore(pesde): pin pesde version

* chore(pesde): pin pesde version to latest

* fix(nix): remove nix lune package from manifest

* fix: wrong path to frktest

* chore(deps): bump `frktest` to `0.0.4`

* chore: update MSLV and package badge link

* ci: fix failures due to bad argument forwarding for `fmt`

* ci: remove `--` to separate script args (https://github.com/lune-org/lune/issues/253)

* chore(deps): bump `asciitable` to `0.2.3`

* chore(LICENSE): undo line ending changes

* chore(scripts): undo line ending changes to scripts

* ci: undo line ending changes to `doc` workflow

* chore(gitignore): undo line ending changes

* fix(tests): pipe buffer overloading

* fix(readme): incorrect term for new require-by-string semantics

* fix(tests): use `dirs` package for temp file creation

* chore(tests): workaround for lune bug without using `sh` pipes

When a command such as `unzip` is spawned, a pipe is established between
lune's stdin and and the `unzip`'s stdout. The pipe buffer has a fixed size
which must not be overflowed, and hence processes yield writing to it, and
wait for the consumer to read the data being written.

`process.create` and `process.exec` wait for the process to exit first, before
calling `read_to_end` on the stream, which works for processes with small outputs,
but for larger ones such as `unzip`'s which involve the process itself waiting for
consumption of data, it creates a cyclic yielding dependence where lune waits for
the process to exit, which wouldn't happen until it finishes writing its data, which
in turn would only happen if it can continue writing chunks after the previous data
has been read (which lune is waiting for the process to exit to begin doing).

We work around this by using a Luau implementation of `ChildReader:readToEnd`.
Specifically, it is backed by a table in order to prevent expensive hashing required
for string concatenation. Another workaround is to compare data as it is being streamed
down to us in real-time, which is only applicable for a few test cases.

* style: apply `stylua` formatting

* chore(tests): address typecheck and lint issues

* chore(zed): update zed configs to reflect vscode configs

---------

Co-authored-by: Erica Marigold <hi@devcomp.xyz>